# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.MenuDataItem:
    properties:
      children:
        items:
          $ref: '#/definitions/dto.MenuDataItem'
        type: array
      icon:
        type: string
      id:
        type: integer
      is_enabled:
        type: boolean
      sort:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
  dto.MenuUpdateRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.MenuDataItem'
        type: array
      type:
        maxLength: 50
        type: string
    required:
    - items
    - type
    type: object
...
  dto.TaskStatusResponse:
    properties:
      message:
        type: string
      progress:
        type: integer
      status:
        type: string
      target:
        type: string
    type: object
  dto.ThemeActivateResponse:
    properties:
      message:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  dto.ThemeInstallResponse:
    properties:
      is_active:
        type: boolean
      message:
        type: string
      name:
        type: string
      slug:
        type: string
      version:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: talen2004@163.com
    name: Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FlecBlog
  version: v1
paths:
...
  /admin/themes:
    get:
      consumes:
      - application/json
      description: 获取所有已注册的主题
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 主题列表
      tags:
      - 主题管理
  /admin/themes/{slug}:
    delete:
      description: 删除指定的非激活主题（磁盘文件 + 数据库记录）
      parameters:
      - description: 主题标识
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除主题
      tags:
      - 主题管理
    get:
      consumes:
      - application/json
      description: 获取指定主题的完整信息
      parameters:
      - description: 主题标识
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 主题详情
      tags:
      - 主题管理
  /admin/themes/{slug}/activate:
    post:
      consumes:
      - application/json
      description: 将指定主题设为激活状态
      parameters:
      - description: 主题标识
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ThemeActivateResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 激活主题
      tags:
      - 主题管理
  /admin/themes/{slug}/config:
    put:
      consumes:
      - application/json
      description: 更新指定主题的配置
      parameters:
      - description: 主题标识
        in: path
        name: slug
        required: true
        type: string
      - description: 主题配置
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新主题配置
      tags:
      - 主题管理
  /admin/themes/{slug}/market-install:
    post:
      consumes:
      - application/json
      description: 从主题市场下载并安装指定主题
      parameters:
      - description: 主题标识
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ThemeInstallResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 从市场安装主题
      tags:
      - 主题管理
  /admin/themes/{slug}/menus:
    get:
      consumes:
      - application/json
      description: 获取指定主题的菜单数据，支持按类型筛选
      parameters:
      - description: 主题标识
        in: path
        name: slug
        required: true
        type: string
      - description: 菜单类型
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuDataItem'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取主题菜单
      tags:
      - 主题管理
    put:
      consumes:
      - application/json
      description: 整体替换指定主题某个类型的菜单列表
      parameters:
      - description: 主题标识
        in: path
        name: slug
        required: true
        type: string
      - description: 菜单数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MenuUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新主题菜单
      tags:
      - 主题管理
  /admin/themes/blog-rebuild-status:
    get:
      consumes:
      - application/json
      description: 主题切换/更新后触发 Blog 重建，前端轮询此接口获取进度
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 查询 Blog 重建状态
      tags:
      - 主题管理
  /admin/themes/install:
    post:
      consumes:
      - multipart/form-data
      description: 从 ZIP 文件安装主题（主题市场内部调用）
      parameters:
      - description: 主题 ZIP 包（≤50MB）
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ThemeInstallResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 安装主题
      tags:
      - 主题管理
  /admin/themes/task-status:
    get:
      consumes:
      - application/json
      description: 获取当前正在进行的主题任务状态（安装/升级/激活/重建）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskStatusResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取任务状态
      tags:
      - 主题管理
...
  /themes/active/schema:
    get:
      consumes:
      - application/json
      description: 获取当前激活主题的 schema 和 config（Blog 端使用）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取激活主题 Schema 和 Config
      tags:
      - 主题
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"